Add python functionality to grab a frame from the camera as an RGB array
Add a cython function py_grab_frame_rgb that accepts a device name and a
width and height resolution, and returns a Numpy ndarray of dimensions
(height, width, 3). Add a test to check that this can be saved as a BMP
file. Clean up the YUV to RGB conversion macros and their coefficients
based on documentation of the YUV and RGB color spaces and conversions
between them. Fix a bug due to the resolution of 640x480 being hard
coded. Add debugging documentation to the README. Stop hard coding
resolution in python test. Fix Gaussian blur.