Add a basic Harris corner detector
Add a basic Harris corner detector that extracts corners based on the properties of the
structure tensor at each point of the input image. Use the Harris response function to
estimate the properties of the eigenvalues of the structure tensor at a point of the
input image. Threshold the value of the Harris response function to determine whether
the point is a corner or not. The edge algorithm implements a rudimentary differential
based edge detection sceheme, rename it to show this. The Canny edge detection algorithm
will be implemented in the future.