Fix edge detection and corner detection
The resulting partial derivatives of an image should be stored as floats instead
of unsigned chars. Also fix differential edge detection so that zero crossings
of the second order directional derivative are detected by looking for a change
of sign in a neighbourhood, rather than checking that its madnitude goes close
to zero. Also add hysteresis thresholding to the differential edge detector.
Implement the Canny Edge Detection algorithm, with non maximum supression and
hysteresis thresholding. Add non maximum supression to the Harris Corner
detector as well.